Add TimerTomato and ui changes.
