+++ src/main/java/ui/timer/ButtonforView.java
@@ -1,4 +1,4 @@
-package ui;
+package ui.timer;
 
 import javax.swing.*;
